[MEDIUM] manage_net: remove incorrect state and poll only once/sec
State 5 was incorrect because it did not set any timeout before
jumping back to state 1 which did the same. So it was remoed.

Also, it is very expensive to check link states at every call,
especially during flashes, and it is not what was performed in
initial version. Let's check links only during state 1.