[OPTIM] add -DQUIET and remove the only unconditional printf
dietlibc suggests to build without *printf to save some space.
Add -DQUIET to remove builtin messages and printf calls. The
only one which was not appropriate to remove was the write to
the pidfile. This one has been replaced with a call to write().

Thanks to the inline attribute added to die(), the code size
has significantly dropped.