[MEDIUM] report status flags to simplify interface management
Instead of having multiple types of interfaces, we now have only one
entry per interface, and a bit field indicating which types of checks
have to be performed, and the status flag indicates their results.

It's up to the functions which need the status to check whether they're
interested in just one bit or the whole status.