[MEDIUM] add support for specific status report via signals
As of now, sending a signal between SIGRTMIN and SIGRTMAX-1 (32..63)
causes the daemon to report some leds combinations (possibly blinking
in a two-cycles mode) for 15s and until one of the monitored interfaces
is plugged, so that it can catch users' attention upon first setup.
The status for each signal is defined with -b <sig> <pattern> when
<pattern> defines two bits per led, one per cycle. This makes it
easy to define on/off/blink together/blink alternatively patterns.

Example :
     alix-leds -i eth0						\
       -b 32 57 -b 33 58 -b 34 56 -b 35 59 -b 36 50 -b 37 61	\
       -l 1 -u -l 2 -d -l 3 -i eth0 -i eth1 -i eth2